AP_Param: signed/unsigned conversion

+++ libraries/AP_Param/tools/eedump_apparam.c
@@ -66,7 +66,7 @@ static uint8_t type_size(enum ap_var_type type)
     case AP_PARAM_MATRIX3F:
         return 3*3*4;
     }
-    printf("unknown type %u\n", type);
+    printf("unknown type %u\n", (unsigned int)type);
     return 0;
 }
